requirements.txt
django==4.0.2
django-environ==0.7.0
Pillow==10.4.0 
djangorestframework==3.15.2
PyJWT==2.8.0
        
